chore: move swapper utils into swapper package
