+++ src/lib/swapper/swappers/OneInchSwapper/utils/oneInchService.ts
@@ -1,6 +1,5 @@
-import { SwapperName } from '@shapeshiftoss/swapper'
+import { createCache, makeSwapperAxiosServiceMonadic, SwapperName } from '@shapeshiftoss/swapper'
 import type { AxiosRequestConfig } from 'axios'
-import { createCache, makeSwapperAxiosServiceMonadic } from 'lib/swapper/utils'
 
 const maxAge = 5 * 1000 // 5 seconds
 // Add cached URLs here if some are cachable
